requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-check==2.6.2
pytest-html==4.2.0
pytest-metadata==3.1.1
python-dotenv==1.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
typing_extensions==4.15.0
urllib3==2.6.3
